pyproject.toml
[project]
name = "demiurg"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.11.11",
    "python-dotenv>=1.0.0",
]

[project.scripts]
demiurg = "demiurg.__main__:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["demiurg"]
